requirements.txt
gtts
moviepy
Pillow